requirements.txt
faiss-cpu
groq
python-dotenv
pypdf
fpdf
PyPDF2
streamlit
langchain
langchain-groq
langchain_google-genai
langchain_community

